CLOUDP-193004: refactor compliancepolicy test structure
